Open link and traverse
The event handler now just opens the first depth of the node. Closes all the open nodes on click.

Test.js and test.html, implements a recursive algorithm to traverse over the linked data structure. It will generate nodes with their parent node tag, so that it makes it easier to navigate the child nodes parent.